# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Redon3 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/checkout:
    post:
      operationId: initiateCheckout
      tags: [billing]
      summary: Initiate Paystack checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL or free plan activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /billing/coupon/validate:
    post:
      operationId: validateCoupon
      tags: [billing]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponRequest"
      responses:
        "200":
          description: Coupon result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponResult"
  /billing/payments:
    get:
      operationId: getPayments
      tags: [billing]
      summary: Get payment history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
  /billing/webhook/paystack:
    post:
      operationId: paystackWebhook
      tags: [billing]
      summary: Paystack webhook receiver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaystackWebhookPayload"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── AUTH SCHEMAS ──────────────────────────────────────────────────────
...
    CheckoutRequest:
      type: object
      properties:
        planId:
          type: string
        couponCode:
          type: ["string", "null"]
        callbackUrl:
          type: string
      required: [planId, callbackUrl]
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: ["string", "null"]
        free:
          type: boolean
        message:
          type: string
      required: [free, message]
    ValidateCouponRequest:
      type: object
      properties:
        code:
          type: string
        planId:
          type: string
      required: [code, planId]
    CouponResult:
      type: object
      properties:
        valid:
          type: boolean
        discountPercent:
          type: number
        originalPriceKobo:
          type: integer
        finalPriceKobo:
          type: integer
        isFree:
          type: boolean
        error:
          type: ["string", "null"]
      required: [valid, discountPercent, originalPriceKobo, finalPriceKobo, isFree]
    Payment:
      type: object
      properties:
        id:
          type: string
        planName:
          type: string
        amountKobo:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, failed]
        type:
          type: string
          enum: [new, renewal, upgrade, coupon_free]
        paystackReference:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, planName, amountKobo, status, type, createdAt]
    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [payments, total, page, limit]

    # ── NOTIFICATION SCHEMAS ──────────────────────────────────────────────
...
    PaystackWebhookPayload:
      type: object
      properties:
        event:
          type: string
        data:
          type: object
      required: [event]
...